+++ extension/api/pom.xml
@@ -11,6 +11,7 @@
 
   <groupId>io.holixon.avro</groupId>
   <artifactId>avro-registry-adapter-api</artifactId>
+  <description>API of the registry and Avro helpers.</description>
 
   <properties>
     <jacoco.skip>false</jacoco.skip>
