+++ extension/registry/apicurio-rest/pom.xml
@@ -11,7 +11,8 @@
   </parent>
 
   <groupId>io.holixon.avro</groupId>
-  <artifactId>avro-registry-adapter-apicurio-rest</artifactId>
+  <artifactId>avro-registry-adapter-apicurio</artifactId>
+  <description>Adapter implementation using Apicurio Schema registry.</description>
 
   <properties>
     <jacoco.skip>false</jacoco.skip>
@@ -20,11 +21,19 @@
   <dependencies>
     <dependency>
       <groupId>io.holixon.avro</groupId>
-      <artifactId>avro-registry-adapter-default</artifactId>
+      <artifactId>avro-registry-adapter-api</artifactId>
     </dependency>
     <dependency>
       <groupId>io.apicurio</groupId>
       <artifactId>apicurio-registry-client</artifactId>
+      <version>${apicurio.version}</version>
     </dependency>
+
+    <dependency>
+      <groupId>io.holixon.avro</groupId>
+      <artifactId>avro-registry-adapter-default</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 </project>
