+++ extension/registry/apicurio-rest/src/main/kotlin/ApicurioAvroSchemaRegistry.kt
@@ -1,4 +1,4 @@
-package io.holixon.avro.adapter.apicurio
+package io.holixon.avro.adapter.registry.apicurio
 
 import io.apicurio.registry.rest.client.RegistryClient
 import io.apicurio.registry.rest.v2.beans.ArtifactMetaData
@@ -8,7 +8,7 @@ import io.apicurio.registry.types.ArtifactType
 import io.holixon.avro.adapter.api.*
 import io.holixon.avro.adapter.api.AvroAdapterApi.byteContent
 import io.holixon.avro.adapter.api.type.AvroSchemaWithIdData
-import io.holixon.avro.adapter.apicurio.AvroAdapterApicurioRest.description
+import io.holixon.avro.adapter.registry.apicurio.AvroAdapterApicurioRest.description
 import org.apache.avro.Schema
 import org.slf4j.LoggerFactory
 import java.io.InputStream
