+++ parent/pom.xml
@@ -13,31 +13,20 @@
   <packaging>pom</packaging>
 
   <properties>
-    <spring-boot.version>2.4.5</spring-boot.version>
     <avro.version>1.10.2</avro.version>
     <kotlin-logging.version>2.0.6</kotlin-logging.version>
-    <jgiven-kotlin.version>1.0.1</jgiven-kotlin.version>
-    <jgiven.version>1.0.0</jgiven.version>
+    <apicurio.version>2.0.0.Final</apicurio.version>
   </properties>
 
   <dependencyManagement>
     <dependencies>
-      <!-- KOTLIN (note: this has to be defined AFTER spring boot or the version is overwritten)-->
+      <!-- KOTLIN -->
       <dependency>
         <groupId>org.jetbrains.kotlin</groupId>
         <artifactId>kotlin-bom</artifactId>
         <version>${kotlin.version}</version>
         <scope>import</scope>
         <type>pom</type>
-      </dependency>
-
-      <!-- SPRING -->
-      <dependency>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-dependencies</artifactId>
-        <version>${spring-boot.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
       </dependency>
 
       <dependency>
